Fixes and reformatting based on student input
The main change here is that <fieldset> should be inside of <form>. A few other small fixes to make it pass validation - e.g. removing the label from the submit button (TIL it's just [not supported](https://www.w3.org/TR/WCAG20-TECHS/H44.html)).
And I also fixed the indentation for the whole file. I've never seen 3 spaces used for indentation before, but that's what you're using, so I stuck with that to make it consistent.