Minor simplifications and indentation fixes
A bit of cleanup for this file following student suggestions. Mostly making whitespace consistent, and getting it to pass validation.
The only actual non-stylistic change is of removing the anonymous identity function in d3.max.
Aaron, was that function ever needed? In an older version of d3 or something? Anyway, if it's no longer relevant, I'll also remove it from the LMS.